# 546/examples/sample-openapi.yaml
openapi: 3.0.0
info:
  title: 用户管理API
  description: 示例OpenAPI规范，用于演示API响应校验工具
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: 开发环境
  - url: https://api.example.com
    description: 生产环境
paths:
  /users:
    get:
      summary: 获取用户列表
      description: 返回所有用户的列表
      responses:
        '200':
          description: 成功返回用户列表
          content:
            application/json:
              schema:
                type: object
                required:
                  - total
                  - data
                properties:
                  total:
                    type: integer
                    description: 总记录数
                    minimum: 0
                  page:
                    type: integer
                    description: 当前页码
                    minimum: 1
                  pageSize:
                    type: integer
                    description: 每页大小
                    minimum: 1
                    maximum: 100
                  data:
                    type: array
                    description: 用户列表
                    items:
                      $ref: '#/components/schemas/User'
    post:
      summary: 创建用户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      summary: 获取单个用户
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: 成功返回用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: 更新用户
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: 删除用户
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 删除成功
  /users/{id}/orders:
    get:
      summary: 获取用户订单
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功返回订单列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
components:
  schemas:
    User:
      type: object
      description: 用户信息
      required:
        - id
        - name
        - email
        - status
      properties:
        id:
          type: integer
          description: 用户ID
          example: 1
        name:
          type: string
          description: 用户姓名
          minLength: 1
          maxLength: 100
          example: 张三
        email:
          type: string
          description: 邮箱地址
          format: email
          example: zhangsan@example.com
        phone:
          type: string
          description: 手机号码
          pattern: '^1[3-9]\d{9}$'
          example: 13800138000
        age:
          type: integer
          description: 年龄
          minimum: 0
          maximum: 150
          example: 25
        status:
          type: string
          description: 用户状态
          enum:
            - active
            - inactive
            - suspended
          example: active
        createdAt:
          type: string
          description: 创建时间
          format: date-time
        updatedAt:
          type: string
          description: 更新时间
          format: date-time
    CreateUserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
        age:
          type: integer
          minimum: 0
          maximum: 150
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
        age:
          type: integer
          minimum: 0
          maximum: 150
        status:
          type: string
          enum:
            - active
            - inactive
            - suspended
    Order:
      type: object
      required:
        - id
        - userId
        - amount
        - status
        - items
      properties:
        id:
          type: string
          description: 订单号
        userId:
          type: integer
          description: 用户ID
        amount:
          type: number
          description: 订单金额
          minimum: 0
        status:
          type: string
          enum:
            - pending
            - paid
            - shipped
            - delivered
            - cancelled
        items:
          type: array
          items:
            type: object
            required:
              - productId
              - quantity
              - price
            properties:
              productId:
                type: integer
              productName:
                type: string
              quantity:
                type: integer
                minimum: 1
              price:
                type: number
                minimum: 0
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: object